requirements.txt
aiobotocore==2.21.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aioitertools==0.12.0
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.8.0
apprise==1.9.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgi-lifespan==2.1.0
asn1crypto==1.5.1
asttokens==3.0.0
async-lru==2.0.4
asyncpg==0.30.0
attrs==25.1.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.3
bigtree==0.25.1
bleach==6.2.0
boto3==1.37.1
botocore==1.37.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
coolname==2.2.0
cryptography==44.0.2
cycler==0.12.1
dask==2025.2.0
dateparser==1.2.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.9
distro==1.9.0
docker==7.1.0
document==1.0
docx==0.2.4
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
graphviz==0.20.3
great-expectations==0.15.45
greenlet==3.1.1
griffe==1.6.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanize==4.12.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.7.4
kiwisolver==1.4.8
load-dotenv==0.1.0
locket==1.0.0
lxml==5.3.1
makefun==1.15.6
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
mmh3==5.1.0
monotonic==1.6
multidict==6.1.0
mypy==1.15.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nc==1.0.2
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
opentelemetry-api==1.30.0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pendulum==3.0.0
pg8000==1.31.2
pillow==11.1.0
platformdirs==4.3.6
posthog==3.19.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==6.30.0
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pyarrow==18.0.0
pycparser==2.22
pydantic
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==3.3.0
python-slugify==8.0.4
python-socks==2.7.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.12
s3transfer==0.11.3
scikit-learn==1.6.1
scipy==1.15.2
scramp==1.4.5
seaborn==0.13.2
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
stack-data==0.6.3
starlette==0.46.1
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.16.0
tinycss2==1.4.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.2
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
ujson==5.10.0
uri-template==1.3.0
urllib3==1.26.20
uv==0.6.6
uvicorn==0.34.0
uvicorn-worker==0.3.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.13
wrapt==1.17.2
xmod==1.8.1
yarl==1.18.3
zipp==3.21.0
reportlab